feat: move many thing back to schemas
