+++ joj/horse/apis/users.py
@@ -20,7 +20,7 @@ router_prefix = "/api/v1"
 @router.get("/{uid}")
 async def get_user(
     user: models.User = Depends(parse_uid),
-) -> StandardResponse[models.UserBase]:
+) -> StandardResponse[schemas.User]:
     return StandardResponse(user)
 
 
@@ -30,7 +30,7 @@ async def list_user_domains(
     ordering: schemas.OrderingQuery = Depends(parse_ordering_query(["name"])),
     pagination: schemas.PaginationQuery = Depends(parse_pagination_query),
     user: models.User = Depends(parse_uid),
-) -> StandardListResponse[models.Domain]:
+) -> StandardListResponse[schemas.Domain]:
     statement = user.find_domains(role)
     domains, count = await models.Domain.execute_list_statement(
         statement, ordering, pagination
@@ -42,7 +42,7 @@ async def list_user_domains(
 async def get_user_problems(
     user: models.User = Depends(parse_uid),
     query: schemas.PaginationQuery = Depends(parse_pagination_query),
-) -> StandardListResponse[models.Problem]:
+) -> StandardListResponse[schemas.Problem]:
     condition = {"owner": user.id}
     cursor = models.Problem.cursor_find(condition, query)
     res = await models.Problem.to_list(cursor)
