+++ joj/horse/models/__init__.py
@@ -1,54 +1,11 @@
 from joj.horse.models.base import BaseORMModel as BaseORMModel
-from joj.horse.models.domain import (
-    Domain as Domain,
-    DomainCreate as DomainCreate,
-    DomainEdit as DomainEdit,
-    DomainTransfer as DomainTransfer,
-)
-from joj.horse.models.domain_invitation import (
-    DomainInvitation as DomainInvitation,
-    DomainInvitationCreate as DomainInvitationCreate,
-    DomainInvitationEdit as DomainInvitationEdit,
-)
-from joj.horse.models.domain_role import (
-    DomainRole as DomainRole,
-    DomainRoleCreate as DomainRoleCreate,
-    DomainRoleEdit as DomainRoleEdit,
-)
-from joj.horse.models.domain_user import (
-    DomainUser as DomainUser,
-    DomainUserAdd as DomainUserAdd,
-    DomainUserPermission as DomainUserPermission,
-    DomainUserUpdate as DomainUserUpdate,
-)
+from joj.horse.models.domain import Domain as Domain
+from joj.horse.models.domain_invitation import DomainInvitation as DomainInvitation
+from joj.horse.models.domain_role import DomainRole as DomainRole
+from joj.horse.models.domain_user import DomainUser as DomainUser
 from joj.horse.models.link_tables import ProblemProblemSetLink as ProblemProblemSetLink
-from joj.horse.models.problem import (
-    Problem as Problem,
-    ProblemClone as ProblemClone,
-    ProblemCreate as ProblemCreate,
-    ProblemEdit as ProblemEdit,
-)
+from joj.horse.models.problem import Problem as Problem
 from joj.horse.models.problem_group import ProblemGroup as ProblemGroup
-from joj.horse.models.problem_set import (
-    ProblemSet as ProblemSet,
-    ProblemSetCreate as ProblemSetCreate,
-    ProblemSetEdit as ProblemSetEdit,
-)
-from joj.horse.models.record import (
-    Record as Record,
-    RecordCaseResult as RecordCaseResult,
-    RecordCodeType as RecordCodeType,
-    RecordResult as RecordResult,
-)
-from joj.horse.models.score import (
-    Score as Score,
-    ScoreBoard as ScoreBoard,
-    UserScore as UserScore,
-)
-from joj.horse.models.user import (
-    User as User,
-    UserBase as UserBase,
-    UserCreate as UserCreate,
-    UserDetail as UserDetail,
-)
+from joj.horse.models.problem_set import ProblemSet as ProblemSet
+from joj.horse.models.user import User as User
 from joj.horse.models.user_oauth_account import UserOAuthAccount as UserOAuthAccount
