+++ joj/horse/schemas/__init__.py
@@ -4,13 +4,64 @@ from joj.horse.schemas.base import (
     StandardListResponse as StandardListResponse,
     StandardResponse as StandardResponse,
 )
+from joj.horse.schemas.domain import (
+    Domain as Domain,
+    DomainCreate as DomainCreate,
+    DomainDetail as DomainDetail,
+    DomainEdit as DomainEdit,
+    DomainTransfer as DomainTransfer,
+)
+from joj.horse.schemas.domain_invitation import (
+    DomainInvitation as DomainInvitation,
+    DomainInvitationCreate as DomainInvitationCreate,
+    DomainInvitationEdit as DomainInvitationEdit,
+)
+from joj.horse.schemas.domain_role import (
+    DomainRole as DomainRole,
+    DomainRoleCreate as DomainRoleCreate,
+    DomainRoleEdit as DomainRoleEdit,
+)
+from joj.horse.schemas.domain_user import (
+    DomainUserAdd as DomainUserAdd,
+    DomainUserPermission as DomainUserPermission,
+    DomainUserUpdate as DomainUserUpdate,
+)
 from joj.horse.schemas.misc import (
     AuthTokens as AuthTokens,
     OAuth2Client as OAuth2Client,
     Redirect as Redirect,
 )
 from joj.horse.schemas.permission import DomainPermission as DomainPermission
+from joj.horse.schemas.problem import (
+    Problem as Problem,
+    ProblemClone as ProblemClone,
+    ProblemCreate as ProblemCreate,
+    ProblemEdit as ProblemEdit,
+)
+from joj.horse.schemas.problem_group import ProblemGroup as ProblemGroup
+from joj.horse.schemas.problem_set import (
+    ProblemSet as ProblemSet,
+    ProblemSetCreate as ProblemSetCreate,
+    ProblemSetEdit as ProblemSetEdit,
+)
 from joj.horse.schemas.query import (
     OrderingQuery as OrderingQuery,
     PaginationQuery as PaginationQuery,
+)
+from joj.horse.schemas.record import (
+    Record as Record,
+    RecordCaseResult as RecordCaseResult,
+    RecordCodeType as RecordCodeType,
+    RecordResult as RecordResult,
+)
+from joj.horse.schemas.score import (
+    Score as Score,
+    ScoreBoard as ScoreBoard,
+    UserScore as UserScore,
+)
+from joj.horse.schemas.user import (
+    User as User,
+    UserCreate as UserCreate,
+    UserDetail as UserDetail,
+    UserWithDomainRole as UserWithDomainRole,
 )
