+++ joj/horse/schemas/problem_group.py
@@ -1,0 +1,14 @@
+from typing import TYPE_CHECKING
+
+from joj.horse.schemas.base import BaseORMSchema, IDMixin, TimestampMixin
+
+if TYPE_CHECKING:
+    pass
+
+
+class ProblemGroup(BaseORMSchema, IDMixin):
+    pass
+
+
+class ProblemGroupDetail(TimestampMixin, ProblemGroup):
+    pass
