+++ joj/horse/schemas/score.py
@@ -1,8 +1,8 @@
 from datetime import datetime, timedelta
 from typing import List
 
-from joj.horse.models.user import UserBase
 from joj.horse.schemas import BaseModel
+from joj.horse.schemas.user import User
 
 
 class Score(BaseModel):
@@ -14,7 +14,7 @@ class Score(BaseModel):
 
 
 class UserScore(BaseModel):
-    user: UserBase
+    user: User
     total_score: int
     total_time_spent: timedelta
     scores: List[Score]
