+++ joj/horse/schemas/user.py
@@ -1,0 +1,83 @@
+import re
+from datetime import datetime
+from typing import TYPE_CHECKING, Any, Dict, Optional, Union
+from uuid import UUID
+
+from pydantic import EmailStr, validator
+from sqlmodel import Field
+
+from joj.horse.models.permission import DefaultRole
+from joj.horse.schemas.base import (
+    BaseORMSchema,
+    IDMixin,
+    SQLModel,
+    TimestampMixin,
+    get_datetime_column,
+    utcnow,
+)
+
+if TYPE_CHECKING:
+    from joj.horse.models import DomainUser
+
+
+UID_RE = re.compile(r"-?\d+")
+UNAME_RE = re.compile(r"[^\s\u3000](.{,254}[^\s\u3000])?")
+
+
+class UserCreate(SQLModel):
+    username: Optional[str] = None
+    email: Optional[str] = None
+    password: Optional[str] = None
+    oauth_name: Optional[str] = None
+    oauth_account_id: Optional[str] = None
+
+    @validator("email", pre=True, always=True)
+    def validate_email(cls, v: Any) -> Optional[EmailStr]:
+        if not v:
+            return None
+        return EmailStr(v)
+
+
+class UserBase(BaseORMSchema):
+    username: str = Field(index=False)
+    email: EmailStr = Field(index=False)
+
+    gravatar: str = Field(default="", index=False)
+    student_id: str = Field(default="")
+    real_name: str = Field(default="")
+    role: str = Field(default=str(DefaultRole.USER))
+    is_active: bool = Field(default=False, index=False)
+
+
+class User(UserBase, IDMixin):
+    pass
+
+
+class UserWithDomainRole(User):
+    domain_id: UUID
+    domain_role: str
+
+    @classmethod
+    def from_domain_user(
+        cls,
+        domain_user: "DomainUser",
+        user: Optional[Union["User", Dict[str, Any]]] = None,
+    ) -> "UserWithDomainRole":
+        if user is None:
+            user = domain_user.user
+        if not isinstance(user, dict):
+            user = user.dict()
+        return cls(
+            **user,  # type: ignore[arg-type]
+            domain_id=domain_user.domain_id,
+            domain_role=domain_user.role,
+        )
+
+
+class UserDetail(TimestampMixin, User):
+    # register_at = fields.DatetimeField(auto_now_add=True)
+    register_ip: str = Field(default="127.0.0.1", index=False)
+    login_at: datetime = Field(
+        sa_column=get_datetime_column(index=False, server_default=utcnow())
+    )
+    login_ip: str = Field(default="127.0.0.1", index=False)
